Перенос методов в иной класс
