+++ 2020-11/spring-02/spring-02-exercise/src/main/java/ru/otus/spring/domain/User.java
@@ -9,12 +9,17 @@ public class User {
     private String name;
     private String surName;
 
-    public User() {
-    }
-
     public void setUser(String name, String surName) {
         this.name = name;
         this.surName = surName;
+    }
+
+    public String getName() {
+        return this.name;
+    }
+
+    public String getSurName() {
+        return this.surName;
     }
 
     public User readUserByConsole() {
