+++ 2020-11/spring-02/spring-02-exercise/src/main/java/ru/otus/spring/service/IQuestionLine.java
@@ -1,7 +1,5 @@
 package ru.otus.spring.service;
 
-import java.util.List;
-
 public interface IQuestionLine {
     String getQuestionText();
     void setQtext(String text);
@@ -14,14 +12,5 @@ public interface IQuestionLine {
 
     public String getCorrectAnswerText();
     void setCorrectAnswerText(String text);
-
-    // Метод подготовки правильных ответов
-    void prepareCorrectAnswers(List<IQuestionLine> qList);
-
-    // Метод проверки ответов на правильность
-    void checkAnswers(List<IQuestionLine> qList);
-
-    // Метод вывода рез-та тестирования
-    void showTestingResult(List<IQuestionLine> qList);
 
 }
