+++ 2020-11/spring-02/spring-02-exercise/src/main/java/ru/otus/spring/service/Questionnaire.java
@@ -1,7 +1,19 @@
 package ru.otus.spring.service;
 
+import ru.otus.spring.domain.User;
+
 import java.io.IOException;
+import java.util.List;
 
 public interface Questionnaire {
     void QuestionnaireExec() throws IOException;
+
+    // Метод подготовки правильных ответов
+    void prepareCorrectAnswers(List<IQuestionLine> qList);
+
+    // Метод проверки ответов на правильность
+    void checkAnswers(List<IQuestionLine> qList);
+
+    // Метод вывода рез-та тестирования
+    void showTestingResult(List<IQuestionLine> qList, User user);
 }
