+++ 2020-11/spring-02/spring-02-exercise/src/main/java/ru/otus/spring/service/QuestionnaireImpl.java
@@ -21,7 +21,6 @@ public class QuestionnaireImpl implements Questionnaire {
 
     @Override
     public void QuestionnaireExec() throws IOException {
-        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext(Main.class);
         // Ввод фамилии и имени пользователя
         User userData = user.readUserByConsole();
 
@@ -29,19 +28,71 @@ public class QuestionnaireImpl implements Questionnaire {
         Scanner scannerConsole = new Scanner(System.in);
         List<IQuestionLine> qList = myFile.readCSVFile();
 
-        IQuestionLine qLine = context.getBean(QuestionLine.class);
-        qLine.prepareCorrectAnswers(qList);
+        prepareCorrectAnswers(qList);
 
         for (int i=0; i < qList.size(); i++) {
             qLine = qList.get(i);
             System.out.print(qLine.getQuestionText());
             qLine.setAnswerText(scannerConsole.nextLine());
         }
-        qLine.checkAnswers(qList);
-        qLine.showTestingResult(qList);
+        checkAnswers(qList);
+        showTestingResult(qList, userData);
 
         // Запуск юнит-теста
         testReadCSVFile();
+    }
+
+    // Метод подготовки правильных ответов
+    public void prepareCorrectAnswers(List<IQuestionLine> qList) {
+        for (int i = 0; i < qList.size(); i++) {
+            IQuestionLine qLine = qList.get(i);
+            String correctAnswerText = "";
+            if (i == 0) {correctAnswerText = "spring";}
+            else if (i == 1) {correctAnswerText = "y";}
+            else if (i == 2) {correctAnswerText = "y";}
+            else if (i == 3) {correctAnswerText = "3";}
+            else if (i == 4) {correctAnswerText = "y";}
+            else {
+                correctAnswerText = "???";
+            }
+            qLine.setCorrectAnswerText(correctAnswerText);
+        }
+
+    }
+
+    // Метод проверки ответов на правильность
+    public void checkAnswers(List<IQuestionLine> qList) {
+        for (int i = 0; i < qList.size(); i++) {
+            IQuestionLine qLine = qList.get(i);
+
+            if (qLine.getAnswerText().equals(qLine.getCorrectAnswerText())) {
+                qLine.setAnswerCorrect(true);
+            }
+            else {
+                qLine.setAnswerCorrect(false);
+            }
+        }
+    }
+
+    // Метод вывода рез-та тестирования
+    public void showTestingResult(List<IQuestionLine> qList, User user) {
+        boolean isWrongAnswerExist = false;
+
+        for (int i = 0; i < qList.size(); i++) {
+            IQuestionLine qLine = qList.get(i);
+
+            if (!qLine.getAnswerCorrect()) {
+                isWrongAnswerExist = true;
+                break;
+            }
+        }
+
+        if (isWrongAnswerExist) {
+            System.out.println("Тестирование студента " + user.getSurName() + " " + user.getName() + " НЕ пройдено. Есть ошибки.");
+        }
+        else {
+            System.out.println("Тестирование студента " + user.getSurName() + " " + user.getName() + " успешно пройдено.");
+        }
     }
 
     @Test
